refactor: extract subscription check into features_service module
Pull HTTP call, caching, and session logic out of
FeatureContentGuard into check_subscription() so
DomainBasedPermission can call it directly
without creating an unsaved model instance.

Same cache keys and TTL.
